requirements.txt
Flask==3.1.0
flask_mail==0.10.0
psycopg2_binary==2.9.10
python-dotenv==1.0.1
python_bcrypt==0.3.2
